refactor(orgtrack): harden session provenance
Pre-commit hook ran. Total eslint: 0, total circular: 0
