+++ docs/frontend-ui-audit-2026-07-14/SessionSidebarReveal.md
@@ -28,17 +28,17 @@
 
 ## D4 — Accessibility
 
-| Line / area                            | Element                                                  | Verdict          | Reason                                                                                                                                                                            | Suggested change |
-| -------------------------------------- | -------------------------------------------------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------- |
-| `NavigationSidebar.tsx` section toggle | `role`, `tabIndex`, `aria-expanded`, Enter/Space handler | keep             | The section can now be expanded/collapsed by keyboard and exposes its state without invalid nested-button markup. Nested header actions are excluded from the parent key handler. | —                |
-| `NavigationMenuRow.tsx` rows           | `data-menu-item-id`, `data-selected`                     | keep with reason | These attributes are stable machine-readable navigation/test state and do not replace the existing semantic/selected styling.                                                     | —                |
-| `NavigationSidebar.tsx` reveal effect  | `scrollIntoView({ block: "nearest" })`                   | keep with reason | It makes the selected historical row visible without moving keyboard focus or forcing a large centered scroll.                                                                    | —                |
+| Line / area                            | Element                                                  | Verdict          | Reason                                                                                                                                                                             | Suggested change |
+| -------------------------------------- | -------------------------------------------------------- | ---------------- | ---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------- |
+| `NavigationSidebar.tsx` section toggle | `role`, `tabIndex`, `aria-expanded`, Enter/Space handler | keep             | The section can now be expanded/collapsed by keyboard and exposes its state without invalid nested-button markup. Nested header actions are excluded from the parent key handler.  | —                |
+| `NavigationMenuRow.tsx` rows           | `data-menu-item-id`, `data-selected`                     | keep with reason | These attributes are stable machine-readable navigation/test state and do not replace the existing semantic/selected styling.                                                      | —                |
+| `NavigationSidebar.tsx` reveal effect  | `scrollIntoView({ block: "nearest" })`                   | keep with reason | It makes the selected historical row visible without moving keyboard focus or forcing a large centered scroll. A monotonic request ID prevents repeat renders from snapping again. | —                |
 
 ## D5 — Visual Patterns Observed
 
-| Pattern                                  | Verdict          | Reason                                                                                                                                                                                                                            |
-| ---------------------------------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
-| Cross-surface “reveal menu row” behavior | keep with reason | The generic sidebar owns scoped scrolling; the workstation connector owns filters, section collapse, and subagent expansion. This avoids embedding workstation rules in Session Blame or duplicating scroll logic at each caller. |
+| Pattern                                  | Verdict          | Reason                                                                                                                                                                                                                                                                                                        |
+| ---------------------------------------- | ---------------- | ------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- |
+| Cross-surface “reveal menu row” behavior | keep with reason | The generic sidebar owns scoped one-shot scrolling; the workstation connector hydrates both the root and child targets and owns filters, section collapse, subagent expansion, and stale-intent cleanup. This avoids embedding workstation rules in Session Blame or duplicating scroll logic at each caller. |
 
 ## Summary
