+++ docs/frontend-ui-audit-2026-07-14/TimelineContent.md
@@ -7,7 +7,7 @@ and existing sidebar conventions manually.
 | Line | Element                                | Verdict          | Reason                                                                                                                                                                                                                      | Suggested change |
 | ---: | -------------------------------------- | ---------------- | --------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- | ---------------- |
 |  196 | Participant row button                 | keep with reason | Native button semantics make each main/subagent row keyboard-focusable. Stable attributes expose root, child target, participant kind, actor, source, and precision to rendered tests without coupling tests to copy.       | None.            |
-|  172 | Action summary                         | keep with reason | All seven canonical resource actions are translated in every shipped locale. Stable read/write count attributes let rendered E2E assert the semantic facts without depending on localized text.                             | None.            |
+|  172 | Action summary                         | keep with reason | All six canonical resource actions are translated in every shipped locale. Stable read/write count attributes let rendered E2E assert the semantic facts without depending on localized text.                               | None.            |
 |  204 | Nested participant spacing             | keep with reason | The participant is indented below its root session while reusing the adjacent sidebar row density and hover token. The border is a single design-system border color, not an arbitrary color.                               | None.            |
 |  219 | Attribution and confidence line        | keep with reason | Localized text distinguishes main-session and subagent facts and states precision explicitly across all 13 shipped locales; truncation preserves the compact sidebar layout and the full value remains in the button title. | None.            |
 |  262 | Root session group                     | keep with reason | The visual hierarchy mirrors the RPC hierarchy: one root session button followed by nested participants. The root remains independently navigable and uses the same session loader as participant rows.                     | None.            |
