+++ src-tauri/crates/agent-cli/Cargo.toml
@@ -23,9 +23,7 @@ rand = { workspace = true }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
 sha2 = "0.10"
+tempfile = "3.13"
 tokio = { workspace = true }
 toml = "0.8"
 tracing = "0.1"
-
-[dev-dependencies]
-tempfile = "3.13"
