+++ src-tauri/crates/orgtrack-core/src/lib.rs
@@ -5,7 +5,6 @@
 //! metadata. Host applications provide paths, permissions, UI commands, and
 //! runtime-specific adapters.
 
-pub mod activity_interaction;
 pub mod canonical;
 pub mod edit_extraction;
 #[cfg(test)]
@@ -15,6 +14,7 @@ pub mod policy;
 pub mod privacy;
 pub mod projectors;
 pub mod repo_sync;
+pub mod resource_interaction;
 pub mod sources;
 pub mod store;
 pub mod sync_export;
