+++ src-tauri/src/orgtrack/session_provenance.rs
@@ -8,135 +8,54 @@
 //! This keeps hooks fast, avoids SQLite contention, and ensures raw prompts,
 //! tool responses, commands, and file contents never enter the spool.
 
-use std::collections::HashMap;
+mod historical_backfill;
+
+pub(super) use historical_backfill::request_historical_backfill;
+
 use std::fs;
 use std::io::Read;
 use std::path::{Component, Path, PathBuf};
 use std::process::Command;
 use std::sync::{Mutex, OnceLock};
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
-use chrono::Utc;
 use core_types::activity::ActivityChunk;
 use core_types::extracted::ExtractedData;
 use core_types::session_event::{EventDisplayStatus, SessionEvent};
 use database::db::get_connection;
-use orgtrack_core::activity_interaction::{
-    activity_chunk_source_event_id, file_interactions_from_activity_chunk,
-    interaction_outcome_from_activity_chunk,
-};
 use orgtrack_core::canonical::{
     AgentMetadata, AttributionPrecision, FileResourceRecord, ResourceAction,
     ResourceInteractionCaptureMethod, ResourceInteractionEnvelopeV1, ResourceInteractionOutcome,
     ResourceInteractionRecord, SessionActorLifecycleEnvelopeV1, SessionActorLifecyclePhase,
     SessionActorRecord, SessionRecord, RESOURCE_INTERACTION_SCHEMA_VERSION,
-    SESSION_ACTOR_SCHEMA_VERSION, SOURCE_ORGII_CLI_SESSIONS, SOURCE_ORGII_RUST_AGENTS,
+    SESSION_ACTOR_SCHEMA_VERSION, SESSION_PROVENANCE_HOOK_ORIGIN,
 };
 use orgtrack_core::hook_adapter::{
     normalize_actor_lifecycle_payload, normalize_hook_payload, HookSource,
 };
 use orgtrack_core::privacy::ORGTRACK_SCHEMA_VERSION;
 use orgtrack_core::repo_sync::paths::{path_hash, record_id};
-use orgtrack_core::sources::claude_code::history::{
-    list_claude_code_history_sessions_paginated, load_claude_code_history_for_session,
+use orgtrack_core::resource_interaction::{
+    activity_chunk_source_event_id, file_interactions_from_activity_chunk,
+    interaction_outcome_from_activity_chunk,
 };
 use orgtrack_core::sources::codex::app::{
-    list_codex_app_sessions_paginated, load_codex_app_for_session, load_codex_app_from_path,
-    resolve_codex_transcript_for_thread_id_near_path,
+    load_codex_app_from_path, resolve_codex_transcript_for_thread_id_near_path,
 };
-use orgtrack_core::sources::cursor_ide::history::{
-    list_cursor_ide_sessions_paginated, load_history_for_session as load_cursor_history_for_session,
-};
-use orgtrack_core::sources::imported_history::cache::{
-    query_cached_sessions_for_repo_from_conn, ImportedHistoryCachedSession,
-};
-use orgtrack_core::sources::imported_history::metadata::{
-    SOURCE_CLAUDE_CODE, SOURCE_CODEX_APP, SOURCE_CURSOR_IDE,
-};
+use orgtrack_core::sources::imported_history::metadata::SOURCE_CODEX_APP;
 use orgtrack_core::store::{sqlite::SqliteRecordStore, RecordStore};
-use rusqlite::Connection;
 use session_persistence::CachedEvent;
 
 const MAX_HOOK_PAYLOAD_BYTES: u64 = 2 * 1024 * 1024;
 const MAX_DRAIN_BATCH: usize = 1_000;
-const HISTORICAL_INTERACTION_PARSER_VERSION: i64 = 1;
-const BACKFILL_REFRESH_INTERVAL: Duration = Duration::from_secs(30);
-const BACKFILL_JOB_RETENTION: Duration = Duration::from_secs(60 * 60);
 static INBOX_DRAIN_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
-static HISTORICAL_BACKFILL_JOBS: OnceLock<Mutex<HashMap<String, HistoricalBackfillJob>>> =
-    OnceLock::new();
-
-#[derive(Debug, Clone)]
-struct HistoricalBackfillJob {
-    status: HistoricalBackfillStatus,
-    indexed_sessions: usize,
-    total_sessions: usize,
-    failed_sessions: usize,
-    last_error: Option<String>,
-    finished_at: Option<Instant>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum HistoricalBackfillStatus {
-    Queued,
-    Discovering,
-    Indexing,
-    Complete,
-    Partial,
-    Failed,
-}
-
-impl HistoricalBackfillStatus {
-    const fn as_str(self) -> &'static str {
-        match self {
-            Self::Queued => "queued",
-            Self::Discovering => "discovering",
-            Self::Indexing => "indexing",
-            Self::Complete => "complete",
-            Self::Partial => "partial",
-            Self::Failed => "failed",
-        }
-    }
-}
-
-impl HistoricalBackfillJob {
-    fn queued() -> Self {
-        Self {
-            status: HistoricalBackfillStatus::Queued,
-            indexed_sessions: 0,
-            total_sessions: 0,
-            failed_sessions: 0,
-            last_error: None,
-            finished_at: None,
-        }
-    }
-
-    fn snapshot(&self) -> super::types::FileSessionHistoryBackfill {
-        super::types::FileSessionHistoryBackfill {
-            status: self.status.as_str().to_string(),
-            indexed_sessions: self.indexed_sessions,
-            total_sessions: self.total_sessions,
-            failed_sessions: self.failed_sessions,
-            last_error: self.last_error.clone(),
-        }
-    }
-
-    fn is_active(&self) -> bool {
-        matches!(
-            self.status,
-            HistoricalBackfillStatus::Queued
-                | HistoricalBackfillStatus::Discovering
-                | HistoricalBackfillStatus::Indexing
-        )
-    }
-}
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub struct ResolvedFileResource {
-    pub repository_id: Option<String>,
-    pub workspace_path: String,
-    pub repo_relative_path: String,
-    pub display_path: String,
+pub(super) struct ResolvedFileResource {
+    pub(super) repository_id: Option<String>,
+    pub(super) workspace_path: String,
+    pub(super) repo_relative_path: String,
+    pub(super) display_path: String,
 }
 
 /// Entry point used by `orgii --session-provenance-hook <source>`.
@@ -249,11 +168,10 @@ fn spool_actor_lifecycle(envelope: &SessionActorLifecycleEnvelopeV1) -> Result<(
 
 /// Drain a bounded number of hook envelopes into the canonical store.
 ///
-/// Malformed/version-incompatible files are quarantined by deletion because
-/// the hook never stores raw source data and the same source event can be
-/// emitted again. Database failures leave the current and remaining files in
-/// place for the next drain.
-pub fn drain_hook_inbox() -> Result<usize, String> {
+/// Malformed/version-incompatible files are moved to a rejected sibling
+/// directory for upgrade diagnostics. Database failures leave the current and
+/// remaining files in place for the next drain.
+pub(super) fn drain_hook_inbox() -> Result<usize, String> {
     let _guard = INBOX_DRAIN_LOCK
         .get_or_init(|| Mutex::new(()))
         .lock()
@@ -297,7 +215,7 @@ pub fn drain_hook_inbox() -> Result<usize, String> {
                 false
             };
         if !persisted {
-            let _ = fs::remove_file(&path);
+            quarantine_invalid_envelope(&inbox, &path)?;
             continue;
         }
         fs::remove_file(&path).map_err(|err| {
@@ -309,6 +227,29 @@ pub fn drain_hook_inbox() -> Result<usize, String> {
         drained += 1;
     }
     Ok(drained)
+}
+
+fn quarantine_invalid_envelope(inbox: &Path, path: &Path) -> Result<(), String> {
+    let rejected = inbox.parent().unwrap_or(inbox).join("rejected");
+    fs::create_dir_all(&rejected)
+        .map_err(|err| format!("Failed to create {}: {err}", rejected.display()))?;
+    let file_name = path
+        .file_name()
+        .ok_or_else(|| format!("Invalid inbox envelope path: {}", path.display()))?;
+    let target = rejected.join(file_name);
+    if target.exists() {
+        // Spool names are content/event identities. Retain the first rejected
+        // copy and discard only an exact-name repeat so the inbox can advance.
+        return fs::remove_file(path)
+            .map_err(|err| format!("Failed to remove {}: {err}", path.display()));
+    }
+    fs::rename(path, &target).map_err(|err| {
+        format!(
+            "Failed to quarantine {} as {}: {err}",
+            path.display(),
+            target.display()
+        )
+    })
 }
 
 fn persist_actor_lifecycle(
@@ -342,7 +283,7 @@ fn persist_actor_lifecycle(
         .or_else(|| codex_rollout_source_session_id(&root_session_id).map(ToString::to_string))
         .unwrap_or_else(|| envelope.source_session_id.clone());
 
-    let root_session = match store.get_session(&root_session_id)? {
+    let mut root_session = match store.get_session(&root_session_id)? {
         Some(mut session) => {
             session.created_at =
                 merge_earliest_timestamp(session.created_at, Some(envelope.occurred_at.as_str()));
@@ -365,11 +306,16 @@ fn persist_actor_lifecycle(
             parent_session_id: None,
             org_member_id: None,
             metadata: AgentMetadata {
-                origin: Some("session_provenance_hook".to_string()),
+                origin: Some(SESSION_PROVENANCE_HOOK_ORIGIN.to_string()),
                 ..AgentMetadata::default()
             },
         },
     };
+    if root_transcript.is_some() {
+        // The resolver found a concrete rollout file. Promote a prior hook
+        // placeholder to a replayable session without waiting for backfill.
+        root_session.metadata.origin = Some(envelope.source.clone());
+    }
     store.upsert_session(&root_session)?;
 
     let transcript = actor_transcript_target(envelope)
@@ -388,6 +334,7 @@ fn persist_actor_lifecycle(
                         Some(envelope.occurred_at.as_str()),
                     );
                 }
+                child.metadata.origin = Some(envelope.source.clone());
                 child
             }
             None => SessionRecord {
@@ -409,7 +356,7 @@ fn persist_actor_lifecycle(
                 parent_session_id: Some(root_session_id.clone()),
                 org_member_id: None,
                 metadata: AgentMetadata {
-                    origin: Some("session_provenance_hook".to_string()),
+                    origin: Some(envelope.source.clone()),
                     display_name: envelope.actor_type.clone(),
                     ..AgentMetadata::default()
                 },
@@ -569,6 +516,9 @@ fn actor_transcript_target(
         return None;
     }
     let path = envelope.transcript_path.as_deref()?.trim();
+    if !Path::new(path).is_file() {
+        return None;
+    }
     let source_session_id = Path::new(path)
         .file_stem()
         .and_then(|value| value.to_str())?
@@ -588,6 +538,9 @@ fn stored_actor_transcript_target(record: &SessionActorRecord) -> Option<(String
         return None;
     }
     let path = record.transcript_path.as_deref()?.trim();
+    if !Path::new(path).is_file() {
+        return None;
+    }
     let source_session_id = Path::new(path)
         .file_stem()
         .and_then(|value| value.to_str())?
@@ -603,7 +556,7 @@ fn stored_actor_transcript_target(record: &SessionActorRecord) -> Option<(String
     ))
 }
 
-pub fn persist_envelope(
+fn persist_envelope(
     store: &dyn RecordStore,
     envelope: &ResourceInteractionEnvelopeV1,
 ) -> Result<(), String> {
@@ -625,7 +578,7 @@ pub fn persist_envelope(
             // be promoted to session-level org membership.
             org_member_id: None,
             metadata: AgentMetadata {
-                origin: Some("session_provenance_hook".to_string()),
+                origin: Some(SESSION_PROVENANCE_HOOK_ORIGIN.to_string()),
                 ..AgentMetadata::default()
             },
         })?;
@@ -652,7 +605,7 @@ pub fn persist_envelope(
 /// Persist the native ORG2 event representation after its tool result has
 /// merged into the tool-call event. This captures reads as well as writes and
 /// preserves exact child-session attribution when the runtime exposes it.
-pub fn persist_native_event_interactions(
+pub(crate) fn persist_native_event_interactions(
     store: &dyn RecordStore,
     session: &SessionRecord,
     event: &SessionEvent,
@@ -752,398 +705,6 @@ pub fn persist_native_event_interactions(
     Ok(())
 }
 
-/// Start (or join) one repository-scoped historical transcript backfill.
-///
-/// The file-history read path only schedules this work. Source discovery and
-/// transcript parsing run on a dedicated blocking thread, so opening a file is
-/// never held hostage by the size of the user's history. Per-session source
-/// fingerprints in SQLite are the durable queue: after a crash or restart the
-/// next request skips current checkpoints and resumes only stale sessions.
-pub fn request_historical_backfill(
-    repo_path: &str,
-    priority_file: &str,
-) -> super::types::FileSessionHistoryBackfill {
-    let canonical_repo = canonicalize_existing_prefix(Path::new(repo_path));
-    let repo_key = canonical_repo.to_string_lossy().into_owned();
-    let jobs = HISTORICAL_BACKFILL_JOBS.get_or_init(|| Mutex::new(HashMap::new()));
-    {
-        let mut jobs = jobs.lock().unwrap_or_else(|poisoned| poisoned.into_inner());
-        jobs.retain(|key, job| {
-            key == &repo_key
-                || job.is_active()
-                || job
-                    .finished_at
-                    .is_some_and(|finished_at| finished_at.elapsed() < BACKFILL_JOB_RETENTION)
-        });
-        if let Some(job) = jobs.get(&repo_key) {
-            let recently_finished = job
-                .finished_at
-                .is_some_and(|finished_at| finished_at.elapsed() < BACKFILL_REFRESH_INTERVAL);
-            if job.is_active() || recently_finished {
-                return job.snapshot();
-            }
-        }
-        jobs.insert(repo_key.clone(), HistoricalBackfillJob::queued());
-    }
-
-    let thread_repo_key = repo_key.clone();
-    let thread_repo_path = canonical_repo.to_string_lossy().into_owned();
-    let priority_file = priority_file.to_string();
-    let spawn_result = std::thread::Builder::new()
-        .name("orgtrack-history-backfill".to_string())
-        .spawn(move || {
-            update_backfill_job(&thread_repo_key, |job| {
-                job.status = HistoricalBackfillStatus::Discovering;
-            });
-            let result = get_connection()
-                .map_err(|err| err.to_string())
-                .and_then(|mut conn| {
-                    reconcile_historical_interactions(
-                        &mut conn,
-                        &thread_repo_path,
-                        &priority_file,
-                        |indexed_sessions, total_sessions, failed_sessions| {
-                            update_backfill_job(&thread_repo_key, |job| {
-                                job.status = HistoricalBackfillStatus::Indexing;
-                                job.indexed_sessions = indexed_sessions;
-                                job.total_sessions = total_sessions;
-                                job.failed_sessions = failed_sessions;
-                            });
-                        },
-                    )
-                });
-            match result {
-                Ok(()) => update_backfill_job(&thread_repo_key, |job| {
-                    job.status = if job.failed_sessions > 0 {
-                        HistoricalBackfillStatus::Partial
-                    } else {
-                        HistoricalBackfillStatus::Complete
-                    };
-                    job.finished_at = Some(Instant::now());
-                }),
-                Err(err) => {
-                    tracing::warn!(
-                        repo_path = %thread_repo_path,
-                        error = %err,
-                        "[SessionProvenance] Historical backfill failed"
-                    );
-                    update_backfill_job(&thread_repo_key, |job| {
-                        job.status = HistoricalBackfillStatus::Failed;
-                        job.last_error = Some(err.clone());
-                        job.finished_at = Some(Instant::now());
-                    });
-                }
-            }
-        });
-    if let Err(err) = spawn_result {
-        update_backfill_job(&repo_key, |job| {
-            job.status = HistoricalBackfillStatus::Failed;
-            job.last_error = Some(format!("Failed to start historical backfill: {err}"));
-            job.finished_at = Some(Instant::now());
-        });
-    }
-
-    let jobs = jobs.lock().unwrap_or_else(|poisoned| poisoned.into_inner());
-    jobs.get(&repo_key)
-        .map(HistoricalBackfillJob::snapshot)
-        .unwrap_or_else(|| HistoricalBackfillJob::queued().snapshot())
-}
-
-fn update_backfill_job(repo_key: &str, update: impl FnOnce(&mut HistoricalBackfillJob)) {
-    let jobs = HISTORICAL_BACKFILL_JOBS.get_or_init(|| Mutex::new(HashMap::new()));
-    let mut jobs = jobs.lock().unwrap_or_else(|poisoned| poisoned.into_inner());
-    update(
-        jobs.entry(repo_key.to_string())
-            .or_insert_with(HistoricalBackfillJob::queued),
-    );
-}
-
-fn reconcile_historical_interactions(
-    conn: &mut Connection,
-    repo_path: &str,
-    priority_file: &str,
-    mut progress: impl FnMut(usize, usize, usize),
-) -> Result<(), String> {
-    let canonical_repo = canonicalize_existing_prefix(Path::new(repo_path));
-    let mut discovery_failures = sync_imported_history_caches(conn);
-
-    let mut imported_batches = Vec::new();
-    for source in [SOURCE_CLAUDE_CODE, SOURCE_CODEX_APP, SOURCE_CURSOR_IDE] {
-        match imported_sessions_for_repo(conn, source, repo_path, &canonical_repo) {
-            Ok(mut sessions) => {
-                sessions.sort_by_key(|session| {
-                    !session_touches_priority_file(session, &canonical_repo, priority_file)
-                });
-                imported_batches.push((source, sessions));
-            }
-            Err(err) => {
-                discovery_failures += 1;
-                tracing::warn!(
-                    source,
-                    error = %err,
-                    "[SessionProvenance] Historical source discovery failed"
-                );
-            }
-        }
-    }
-    let native_sessions = native_sessions_for_repo(conn, repo_path, &canonical_repo)?;
-    let total_sessions = imported_batches
-        .iter()
-        .map(|(_, sessions)| sessions.len())
-        .sum::<usize>()
-        + native_sessions.len()
-        + discovery_failures;
-    let mut indexed_sessions = 0;
-    let mut failed_sessions = discovery_failures;
-    progress(indexed_sessions, total_sessions, failed_sessions);
-
-    for (source, sessions) in imported_batches {
-        for session in sessions {
-            match reconcile_imported_session(conn, source, &canonical_repo, &session) {
-                Ok(()) => indexed_sessions += 1,
-                Err(err) => {
-                    failed_sessions += 1;
-                    tracing::warn!(
-                        source,
-                        session_id = %session.session_id,
-                        error = %err,
-                        "[SessionProvenance] Historical session reconciliation failed"
-                    );
-                }
-            }
-            progress(indexed_sessions, total_sessions, failed_sessions);
-        }
-    }
-    for session in native_sessions {
-        match reconcile_native_session(conn, &canonical_repo, &session) {
-            Ok(()) => indexed_sessions += 1,
-            Err(err) => {
-                failed_sessions += 1;
-                tracing::warn!(
-                    session_id = %session.session_id,
-                    error = %err,
-                    "[SessionProvenance] Native historical session reconciliation failed"
-                );
-            }
-        }
-        progress(indexed_sessions, total_sessions, failed_sessions);
-    }
-    Ok(())
-}
-
-fn sync_imported_history_caches(conn: &mut Connection) -> usize {
-    let mut failures = 0;
-    if let Err(err) = list_claude_code_history_sessions_paginated(conn, 1, 0) {
-        failures += 1;
-        tracing::warn!(error = %err, "[SessionProvenance] Claude history discovery failed");
-    }
-    if let Err(err) = list_codex_app_sessions_paginated(conn, 1, 0) {
-        failures += 1;
-        tracing::warn!(error = %err, "[SessionProvenance] Codex history discovery failed");
-    }
-    if let Err(err) = list_cursor_ide_sessions_paginated(conn, 1, 0) {
-        failures += 1;
-        tracing::warn!(error = %err, "[SessionProvenance] Cursor history discovery failed");
-    }
-    failures
-}
-
-fn imported_sessions_for_repo(
-    conn: &Connection,
-    source: &str,
-    repo_path: &str,
-    canonical_repo: &Path,
-) -> Result<Vec<ImportedHistoryCachedSession>, String> {
-    let canonical_repo_str = canonical_repo.to_string_lossy();
-    let mut sessions_by_id = std::collections::BTreeMap::new();
-    for candidate in [repo_path, canonical_repo_str.as_ref()] {
-        for session in query_cached_sessions_for_repo_from_conn(conn, source, candidate)? {
-            sessions_by_id.insert(session.session_id.clone(), session);
-        }
-    }
-    Ok(sessions_by_id.into_values().collect())
-}
-
-fn session_touches_priority_file(
-    session: &ImportedHistoryCachedSession,
-    canonical_repo: &Path,
-    priority_file: &str,
-) -> bool {
-    let priority_file = priority_file
-        .trim_start_matches(['/', '\\'])
-        .replace('\\', "/");
-    session.impact.touched_files.iter().any(|candidate| {
-        let candidate_path = Path::new(candidate);
-        let relative = candidate_path
-            .strip_prefix(canonical_repo)
-            .unwrap_or(candidate_path)
-            .to_string_lossy()
-            .trim_start_matches(['/', '\\'])
-            .replace('\\', "/");
-        relative == priority_file
-    })
-}
-
-fn reconcile_imported_session(
-    conn: &Connection,
-    source: &str,
-    canonical_repo: &Path,
-    session: &ImportedHistoryCachedSession,
-) -> Result<(), String> {
-    let store = SqliteRecordStore::new(conn);
-    let fingerprint = imported_session_fingerprint(session);
-    if store.interaction_import_is_current(
-        source,
-        &session.session_id,
-        &fingerprint,
-        HISTORICAL_INTERACTION_PARSER_VERSION,
-    )? {
-        return Ok(());
-    }
-    let chunks = match source {
-        SOURCE_CLAUDE_CODE => load_claude_code_history_for_session(conn, &session.session_id),
-        SOURCE_CODEX_APP => load_codex_app_for_session(conn, &session.session_id),
-        SOURCE_CURSOR_IDE => load_cursor_history_for_session(&session.session_id),
-        _ => return Err(format!("Unsupported imported history source: {source}")),
-    }?;
-    store.delete_reconciled_resource_interactions(source, &session.session_id)?;
-    let actor_id = session
-        .parent_session_id
-        .as_ref()
-        .map(|_| session.source_session_id.as_str());
-    let precision = if actor_id.is_some() {
-        AttributionPrecision::Exact
-    } else {
-        AttributionPrecision::SessionOnly
-    };
-    persist_activity_chunks(
-        &store,
-        source,
-        Some(&session.source_session_id),
-        &session.session_id,
-        actor_id,
-        session
-            .repo_path
-            .as_deref()
-            .or_else(|| canonical_repo.to_str())
-            .unwrap_or("."),
-        precision,
-        &chunks,
-    )?;
-    store.mark_interaction_imported(
-        source,
-        &session.session_id,
-        &fingerprint,
-        HISTORICAL_INTERACTION_PARSER_VERSION,
-        &Utc::now().to_rfc3339(),
-    )
-}
-
-fn imported_session_fingerprint(session: &ImportedHistoryCachedSession) -> String {
-    format!(
-        "{}:{}:{}:{}",
-        session.source_mtime_ms,
-        session.source_size_bytes,
-        session.source_fingerprint,
-        session.parser_version
-    )
-}
-
-fn native_sessions_for_repo(
-    conn: &Connection,
-    repo_path: &str,
-    canonical_repo: &Path,
-) -> Result<Vec<SessionRecord>, String> {
-    let store = SqliteRecordStore::new(conn);
-    let canonical_repo_str = canonical_repo.to_string_lossy();
-    let mut sessions_by_id = std::collections::BTreeMap::new();
-    for candidate in [repo_path, canonical_repo_str.as_ref()] {
-        for session in store.list_sessions(Some(candidate))? {
-            sessions_by_id.insert(session.session_id.clone(), session);
-        }
-    }
-    Ok(sessions_by_id
-        .into_values()
-        .filter(|session| {
-            matches!(
-                session.source.as_str(),
-                SOURCE_ORGII_RUST_AGENTS | SOURCE_ORGII_CLI_SESSIONS
-            ) && session.workspace_path.as_deref().is_some_and(|workspace| {
-                canonicalize_existing_prefix(Path::new(workspace)) == canonical_repo
-            })
-        })
-        .collect())
-}
-
-fn reconcile_native_session(
-    conn: &Connection,
-    canonical_repo: &Path,
-    session: &SessionRecord,
-) -> Result<(), String> {
-    let Some(metadata) = session_persistence::get_session_metadata(&session.session_id)
-        .map_err(|err| err.to_string())?
-    else {
-        // A live-only session may not have a persisted event cache. It has
-        // still been fully considered for coverage and is not a failure.
-        return Ok(());
-    };
-    let fingerprint = format!(
-        "{}:{}:{}:{}",
-        metadata.event_count,
-        metadata.cached_at,
-        metadata.time_range_start.as_deref().unwrap_or_default(),
-        metadata.time_range_end.as_deref().unwrap_or_default()
-    );
-    let store = SqliteRecordStore::new(conn);
-    if store.interaction_import_is_current(
-        &session.source,
-        &session.session_id,
-        &fingerprint,
-        HISTORICAL_INTERACTION_PARSER_VERSION,
-    )? {
-        return Ok(());
-    }
-    let events =
-        session_persistence::load_events(&session.session_id).map_err(|err| err.to_string())?;
-    let chunks = events
-        .iter()
-        .map(cached_event_to_activity_chunk)
-        .collect::<Vec<_>>();
-    let actor_id = session.org_member_id.as_deref().or_else(|| {
-        session
-            .parent_session_id
-            .as_ref()
-            .map(|_| session.source_session_id.as_str())
-    });
-    let precision = if actor_id.is_some() {
-        AttributionPrecision::Exact
-    } else {
-        AttributionPrecision::SessionOnly
-    };
-    store.delete_reconciled_resource_interactions(&session.source, &session.session_id)?;
-    persist_activity_chunks(
-        &store,
-        &session.source,
-        Some(&session.source_session_id),
-        &session.session_id,
-        actor_id,
-        session
-            .workspace_path
-            .as_deref()
-            .or_else(|| canonical_repo.to_str())
-            .unwrap_or("."),
-        precision,
-        &chunks,
-    )?;
-    store.mark_interaction_imported(
-        &session.source,
-        &session.session_id,
-        &fingerprint,
-        HISTORICAL_INTERACTION_PARSER_VERSION,
-        &Utc::now().to_rfc3339(),
-    )
-}
-
 #[allow(clippy::too_many_arguments)]
 fn persist_activity_chunks(
     store: &dyn RecordStore,
@@ -1235,15 +796,16 @@ fn persist_file_interaction(
         path_hash: path_hash(&resolved.repo_relative_path),
     })?;
 
-    let interaction_id = record_id(&[
-        "interaction",
+    let interaction_id = resource_interaction_id(
         source,
         session_id,
-        source_event_id.unwrap_or(""),
+        source_event_id,
+        actor_id,
         &resource_id,
-        action.as_str(),
+        action,
         occurred_at,
-    ]);
+        capture_method,
+    );
     store.append_resource_interaction(&ResourceInteractionRecord {
         schema_version: RESOURCE_INTERACTION_SCHEMA_VERSION,
         interaction_id,
@@ -1262,7 +824,34 @@ fn persist_file_interaction(
     })
 }
 
-pub fn resolve_file_resource(cwd: &str, file_path: &str) -> ResolvedFileResource {
+#[allow(clippy::too_many_arguments)]
+fn resource_interaction_id(
+    source: &str,
+    session_id: &str,
+    source_event_id: Option<&str>,
+    actor_id: Option<&str>,
+    resource_id: &str,
+    action: ResourceAction,
+    occurred_at: &str,
+    capture_method: ResourceInteractionCaptureMethod,
+) -> String {
+    // A hook observation and a later reconciled observation are distinct
+    // immutable facts even when they describe the same vendor tool event.
+    // The read projection correlates and selects the strongest one.
+    record_id(&[
+        "interaction",
+        source,
+        session_id,
+        source_event_id.unwrap_or(""),
+        actor_id.unwrap_or(""),
+        resource_id,
+        action.as_str(),
+        occurred_at,
+        capture_method.as_str(),
+    ])
+}
+
+pub(super) fn resolve_file_resource(cwd: &str, file_path: &str) -> ResolvedFileResource {
     // Resolve aliases such as macOS `/tmp` -> `/private/tmp` on both sides
     // before comparing them. For create/delete events the leaf may not exist,
     // so canonicalize the longest existing prefix and reattach the tail.
@@ -1300,7 +889,7 @@ pub fn resolve_file_resource(cwd: &str, file_path: &str) -> ResolvedFileResource
 /// Periodically drains external hook envelopes while the desktop app is open.
 /// A bounded drain keeps each blocking task predictable; remaining envelopes
 /// are retried on the next tick and on file-history queries.
-pub fn spawn_hook_inbox_drain_loop() {
+pub(crate) fn spawn_hook_inbox_drain_loop() {
     tauri::async_runtime::spawn(async {
         loop {
             let result = tauri::async_runtime::spawn_blocking(drain_hook_inbox).await;
@@ -1380,6 +969,48 @@ fn canonicalize_existing_prefix(path: &Path) -> PathBuf {
 #[cfg(test)]
 mod tests {
     use super::*;
+    use orgtrack_core::sources::codex::app::load_codex_app_for_session;
+    use orgtrack_core::sources::imported_history::metadata::SOURCE_CLAUDE_CODE;
+    use rusqlite::Connection;
+
+    #[test]
+    fn interaction_identity_preserves_independent_capture_observations() {
+        let hook_id = resource_interaction_id(
+            SOURCE_CLAUDE_CODE,
+            "root-session",
+            Some("tool-1"),
+            None,
+            "file-1",
+            ResourceAction::Read,
+            "2026-07-14T01:00:00Z",
+            ResourceInteractionCaptureMethod::Hook,
+        );
+        let reconciled_id = resource_interaction_id(
+            SOURCE_CLAUDE_CODE,
+            "child-session",
+            Some("tool-1"),
+            Some("agent-1"),
+            "file-1",
+            ResourceAction::Read,
+            "2026-07-14T01:00:00Z",
+            ResourceInteractionCaptureMethod::Reconciled,
+        );
+
+        assert_ne!(hook_id, reconciled_id);
+        assert_eq!(
+            hook_id,
+            resource_interaction_id(
+                SOURCE_CLAUDE_CODE,
+                "root-session",
+                Some("tool-1"),
+                None,
+                "file-1",
+                ResourceAction::Read,
+                "2026-07-14T01:00:00Z",
+                ResourceInteractionCaptureMethod::Hook,
+            )
+        );
+    }
 
     #[test]
     fn resolves_relative_paths_without_leaking_content() {
@@ -1394,6 +1025,20 @@ mod tests {
         );
         assert_eq!(resolved.repo_relative_path, "src/lib.rs");
         assert_eq!(resolved.display_path, "src/lib.rs");
+    }
+
+    #[test]
+    fn malformed_envelopes_are_quarantined_for_upgrade_diagnostics() {
+        let temp = tempfile::tempdir().expect("temporary provenance root");
+        let inbox = temp.path().join("inbox");
+        fs::create_dir_all(&inbox).expect("inbox");
+        let path = inbox.join("invalid.json");
+        fs::write(&path, b"not-json").expect("invalid envelope");
+
+        quarantine_invalid_envelope(&inbox, &path).expect("quarantine invalid envelope");
+
+        assert!(!path.exists());
+        assert!(temp.path().join("rejected").join("invalid.json").is_file());
     }
 
     #[test]
@@ -1502,6 +1147,8 @@ mod tests {
             .expect("parent session");
         assert_eq!(parent.created_at.as_deref(), Some("2026-07-14T01:00:00Z"));
         assert_eq!(parent.updated_at.as_deref(), Some("2026-07-14T01:02:00Z"));
+        assert_eq!(parent.metadata.origin.as_deref(), Some(SOURCE_CODEX_APP));
+        assert_eq!(child.metadata.origin.as_deref(), Some(SOURCE_CODEX_APP));
         let parent_chunks = load_codex_app_for_session(&conn, &parent_session_id)
             .expect("load parent through lifecycle locator");
         assert!(parent_chunks.iter().any(|chunk| {
