+++ src/api/tauri/lineage/index.ts
@@ -15,7 +15,6 @@ import type {
   OrgtrackExportResult,
   OrgtrackExtractionMemoryGate,
   OrgtrackFileSessionHistory,
-  OrgtrackFileSessionLookup,
   OrgtrackFileTimeline,
   OrgtrackIndex,
   OrgtrackSessionCheckpoint,
@@ -36,7 +35,6 @@ export type {
   OrgtrackExportResult,
   OrgtrackExtractionMemoryGate,
   OrgtrackCommitLink,
-  OrgtrackFileSessionLookup,
   OrgtrackFileSessionHistory,
   OrgtrackFileTimeline,
   OrgtrackIndex,
