+++ src/i18n/locales/ko/common.json
@@ -408,15 +408,13 @@
     "sessionBlame": "세션 귀속",
     "sessionBlameSubagent": "하위 에이전트: {{name}}",
     "sessionBlameMainSession": "기본 세션",
-    "sessionBlameSessionAttribution": "세션 수준 귀속",
     "sessionBlameAction": {
       "read": "읽기",
       "write": "쓰기",
       "create": "생성",
       "delete": "삭제",
       "rename": "이름 변경",
-      "search": "검색",
-      "reference": "참조"
+      "search": "검색"
     },
     "sessionBlamePrecision": {
       "unknown": "신뢰도 알 수 없음",
