+++ src/store/session/sessionAtom/__tests__/sidebarLoaders.test.ts
@@ -6,6 +6,7 @@ import { IMPORTED_HISTORY_SOURCES } from "@src/api/tauri/externalHistory";
 import { dataSourceConfigAtom } from "../../dataSourceConfigAtom";
 import { sessionsAtom } from "../atoms";
 import {
+  __TESTS_ONLY,
   loadMoreCategory,
   loadSidebarSessionById,
   loadSidebarSessions,
@@ -206,7 +207,57 @@ describe("loadSidebarSessions", () => {
         limit: 1,
       })
     );
+    expect(mocks.sessionAggregateList.mock.calls[0]?.[0]).not.toHaveProperty(
+      "disabledExternalHistorySources"
+    );
     expect(mocks.externalHistorySidebarList).not.toHaveBeenCalled();
     expect(mocks.store?.get(sessionsAtom)).toContainEqual(historicalSession);
+  });
+
+  it("does not erase an exact-loaded child during a provider first-page refresh", () => {
+    const codex = IMPORTED_HISTORY_SOURCES.find(
+      (source) => source.sourceId === "codex_app"
+    );
+    expect(codex).toBeTruthy();
+    if (!codex) return;
+
+    const oldRoot = {
+      session_id: "codexapp-old-root",
+      name: "Old root",
+      status: "completed" as const,
+      created_at: "2026-07-01T00:00:00Z",
+      updated_at: "2026-07-01T00:00:00Z",
+    };
+    const exactChild = {
+      ...oldRoot,
+      session_id: "codexapp-exact-child",
+      name: "Exact child",
+      parentSessionId: "codexapp-current-root",
+    };
+    const currentRoot = {
+      ...oldRoot,
+      session_id: "codexapp-current-root",
+      name: "Current root",
+      updated_at: "2026-07-14T00:00:00Z",
+    };
+
+    const replaced = __TESTS_ONLY.replaceExternalHistorySourceFirstPage(
+      [oldRoot, exactChild],
+      [currentRoot],
+      codex
+    );
+
+    expect(replaced.map((session) => session.session_id)).toEqual([
+      "codexapp-current-root",
+      "codexapp-exact-child",
+    ]);
+
+    const disabled = __TESTS_ONLY.replaceExternalHistorySourceFirstPage(
+      replaced,
+      [],
+      codex,
+      false
+    );
+    expect(disabled).toEqual([]);
   });
 });
