+++ src/store/ui/__tests__/sidebarAtom.test.ts
@@ -1,6 +1,7 @@
 import { createStore } from "jotai/vanilla";
 
 import {
+  clearSessionSidebarRevealAtom,
   requestSessionSidebarRevealAtom,
   sessionSidebarRevealRequestAtom,
 } from "../sidebarAtom";
@@ -32,5 +33,28 @@ describe("requestSessionSidebarRevealAtom", () => {
     store.set(requestSessionSidebarRevealAtom, { sessionId: "   " });
 
     expect(store.get(sessionSidebarRevealRequestAtom)).toBeNull();
+  });
+
+  it("clears only the reveal request that was actually completed", () => {
+    const store = createStore();
+    store.set(requestSessionSidebarRevealAtom, { sessionId: "session-a" });
+    const firstRequestId = store.get(
+      sessionSidebarRevealRequestAtom
+    )!.requestId;
+    store.set(requestSessionSidebarRevealAtom, { sessionId: "session-b" });
+
+    store.set(clearSessionSidebarRevealAtom, firstRequestId);
+    expect(store.get(sessionSidebarRevealRequestAtom)?.sessionId).toBe(
+      "session-b"
+    );
+
+    store.set(
+      clearSessionSidebarRevealAtom,
+      store.get(sessionSidebarRevealRequestAtom)!.requestId
+    );
+    expect(store.get(sessionSidebarRevealRequestAtom)).toBeNull();
+
+    store.set(requestSessionSidebarRevealAtom, { sessionId: "session-c" });
+    expect(store.get(sessionSidebarRevealRequestAtom)?.requestId).toBe(3);
   });
 });
